Añadido API de asistencias Refactorizacion de transformers
