+++ app/Transformer/Goals/GoalPlayerTransformer.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Transformer\Goals;
+
+class GoalPlayerTransformer {
+
+    public function transform($player){
+        return [$player];
+    }
+}
