+++ app/Transformer/Goals/GoalTeamTransformer.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Transformer\Goals;
+
+class GoalTeamTransformer {
+
+    public function transform($team){
+        return [$team];
+    }
+}
