+++ app/Transformer/Matches/MatchRoundTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Transformer;
+namespace App\Transformer\Matches;
 
 class MatchRoundTransformer {
