+++ app/Transformer/Matches/MatchTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Transformer;
+namespace App\Transformer\Matches;
 
 class MatchTransformer {
