+++ app/Transformer/Players/PlayerTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Transformer;
+namespace App\Transformer\Players;
 
 class PlayerTransformer {
