+++ app/Transformer/Rounds/RoundTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Transformer;
+namespace App\Transformer\Rounds;
 
 class RoundTransformer {
