+++ app/Transformer/Seasons/SeasonRoundsTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Transformer;
+namespace App\Transformer\Seasons;
 
 class SeasonRoundsTransformer {
