+++ app/Transformer/Seasons/SeasonTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Transformer;
+namespace App\Transformer\Seasons;
 
 class SeasonTransformer {
