+++ app/Transformer/Teams/TeamPlayerTransformer.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Transformer;
+namespace App\Transformer\Teams;
 
-class SeasonTeamsTransformer {
+class TeamPlayerTransformer {
 
     public function transform($teams){
         return [$teams];
